set hyper 1 as new default
